# 3.spec-kit-openspec/java/specs/001-multi-level-categories/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CMS API - Multi-Level Categories & Rich Text Enhancement
  description: API specification for CMS with multi-level categories, rich text editor, and image upload
  version: 1.0.0
  contact:
    name: CMS Development Team
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.cms.example.com/api
    description: Production server
...
paths:
...
  /images/upload:
    post:
      tags:
        - Images
      summary: Upload an image
      description: |
        Uploads an image file. Client should compress before upload.

        Supported formats: JPEG, PNG, GIF, WebP
        Max file size: 10MB (recommend compress to 2MB on client)

        Security: Server validates magic bytes, re-encodes image, generates safe filename.
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
              required:
                - file
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadResponse'
        '400':
          description: Invalid file (wrong type, too large, corrupted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File size exceeds 10MB limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /images/{id}:
    get:
      tags:
        - Images
      summary: Get image metadata
      description: Retrieves image metadata (not the file itself).
      operationId: getImageById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Images
      summary: Delete an image
      description: Deletes an image. Warning - may break content that references this image.
      operationId: deleteImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Image deleted successfully
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        filename:
          type: string
          description: Server-generated UUID filename
          example: "550e8400-e29b-41d4-a716-446655440000.jpg"
        originalFilename:
          type: string
          example: "vue-logo.png"
        mimeType:
          type: string
          enum: ["image/jpeg", "image/png", "image/gif", "image/webp"]
          example: "image/png"
        fileSize:
          type: integer
          description: File size in bytes
          example: 245760
        width:
          type: integer
          description: Image width in pixels
          example: 1920
        height:
          type: integer
          description: Image height in pixels
          example: 1080
        url:
          type: string
          description: Public URL path for embedding in content
          example: "/uploads/550e8400-e29b-41d4-a716-446655440000.jpg"
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - filename
        - originalFilename
        - mimeType
        - fileSize
        - url
        - uploadedAt
    ImageUploadResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        url:
          type: string
          description: URL to use in content img src
          example: "/uploads/550e8400-e29b-41d4-a716-446655440000.jpg"
        filename:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000.jpg"
        originalFilename:
          type: string
          example: "vue-logo.png"
        mimeType:
          type: string
          example: "image/png"
        fileSize:
          type: integer
          example: 245760
        width:
          type: integer
          example: 1920
        height:
          type: integer
          example: 1080
      required:
        - id
        - url
        - filename
        - originalFilename
        - mimeType
        - fileSize

    # ==========================================
    # Common Schemas
    # ==========================================
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Category name 'Technology' already exists in this parent scope"
        path:
          type: string
          example: "/api/categories"
        validationErrors:
          type: array
          description: Field-level validation errors (optional)
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - timestamp
        - status
        - error
        - message
        - path